requirements.txt
# API и сервер
fastapi==0.110.0
uvicorn==0.27.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# База данных и метаданные
psycopg2-binary==2.9.9
sqlalchemy==2.0.28  # Пригодится для удобной работы с типами, даже если используем чистый SQL

# ML и Векторный поиск
chromadb==0.4.24
sentence-transformers==2.5.1
openai==1.13.3  # Клиент для OpenRouter

# Агентская логика
langgraph==0.0.26
langchain==0.1.11
langchain-openai==0.0.8